Fix segmentation tool bugs
