+++ packages/tools/src/tools/segmentation/strategies/utils/stackVolumeCheck.ts
@@ -1,0 +1,19 @@
+import {
+  SegToolsEditData,
+  SegToolsEditDataStack,
+  SegToolsEditDataVolume,
+} from '../../../../types';
+
+function isStackSegmentation(
+  editData: SegToolsEditData
+): editData is SegToolsEditDataStack {
+  return (editData as SegToolsEditDataStack).imageIds !== undefined;
+}
+
+function isVolumeSegmentation(
+  editData: SegToolsEditData
+): editData is SegToolsEditDataVolume {
+  return (editData as SegToolsEditDataVolume).imageVolume !== undefined;
+}
+
+export { isStackSegmentation, isVolumeSegmentation };
